Add ability to pick which char hides lines in cfg
We still need to deal with the rust-specific hide-line code that checks
for the `!` and `[` chars to determine if lines should be hidden or not.
Ideally, this code shouldn't be here at all. Rust should just pick a
hide char that doesn't conflict with the language. But for the sake of
backwards compatbility, this should be dealt with appropriately (for
now at least)

Also, if the user picks a hide character that can be interpretted by
regex, this can cause the code to panic and fail. While this isn't
ideal, it doesn't necessarily seem like something the code should have
to accommodate. Users should just be noted in the docs that they need to
use the `\\` to escape special regex chars if they plan to use them.
We could do better with error handling though >_<